Adding the username and id to the jwt token.
It is helpful to have the user id included in the token to give the angular app a way to identify the user directly. The username is more of a convenience thing, it is included in the header. This way we don't need to immediately request a user object after logging in.