Improving the application flow with validation.
Changed the Save And Continue Behavior to validate & notify the user of any problems prior to advancing.
When advancing via the ‘Save And Continue button, it only validates the current page.
Fixed a bug with required fields nested within a line item with an ask_if condition. The ask_if condition was being ignored.