Correcting a bug with validating sub-workflows.
If a workflow (or single line item) is marked as required and is empty, it would get it's validation results set to 'This field is required'. This is problematic when validation is applied to sub-elements, because then failures would try to get written as properties of the string value. This results in cryptic 'unable to write to read-only properties' errors. Also, added more comments to the validation method.